32 Bit image compression
Move both FiFOs to 32b because having both FiFOs work on different
clocks were causing us issues.

The downside to this is that there is no VGA support, use a different
project file to get the lighting right.

32 bits are stored 16 in FiFO1 and 16 in FiFo2, and will be read out
into 32 HPS PIOs for the HPS to extract into the image array.