ROI FSM 3/4 Done
Found ROI TOP, and a little bit of LEFT and RIGHT

LEFT and RIGHT are found right after TOP is found.
LEFT looks for white then black pixels in a row. Once this is found, the
FSM looks for black then white pixels to denote LEFt and RIGHT

The C Code is also changed to display the HW ROI indices alongside the
SW calculated indices